ft(404Page): factor out common navbar links
 - factor out common navbar links

[Finishes #167566810]
