+++ puzzled_ui/components/commons/navbarContainer.tsx
@@ -1,31 +1,16 @@
 import * as React from "react";
 import { NavBar } from './navBar'
 import { AppName } from './appName'
-import { navbarInterface } from "../interfaces";
 
-let linkNoStyle: string = 'link__no-style';
-
-let navBarLinks: navbarInterface = {
-    primaryLabel: { text: 'puzzled', href: '/', style: linkNoStyle },
-    secLabel: { text: null, href: '#' },
-    links:
-        [
-            { name: 'Games', href: '#' },
-            { name: 'Pin', href: '#' },
-            { name: null, href: null, linkClass: 'navbar-separator' },
-            { name: 'Sign In', href: '#' },
-            { name: 'Sign Up', href: '#' }
-
-        ]
-};
+import { defaultNavBarLinks } from "./navbarLinks";
 
 function NavBarContainer(): JSX.Element {
     return (
         <div className='navbar-container'>
             <NavBar
-                secLabel={ navBarLinks.secLabel }
-                primaryLabel={ navBarLinks.primaryLabel }
-                links={ navBarLinks.links } key={ 'main' }
+                secLabel={ defaultNavBarLinks.secLabel }
+                primaryLabel={ defaultNavBarLinks.primaryLabel }
+                links={ defaultNavBarLinks.links } key={ 'main' }
             />
             <AppName/>
         </div>
