+++ puzzled_ui/components/sudoku/sudoku.tsx
@@ -1,24 +1,7 @@
 import * as React from "react";
 import { SudokuGrid, defaultSudokuType } from './sudokuGrid'
 import { NavBar as SudokuNavBar }  from '../commons/NavBar';
-import { navbarInterface } from "../interfaces";
-
-let linkNoStyle: string = 'link__no-style';
-
-let sudokuNavBarLinks: navbarInterface = {
-    primaryLabel: { text: 'puzzled', href: "/", style: linkNoStyle },
-    secLabel: { text: 'sudoku', href: '/sudoku/', style: linkNoStyle },
-    links:
-        [
-            { name: 'Play', href: '#' },
-            { name: 'Solve', href: '#' },
-            { name: 'About Sudoku', href: '#' },
-            { name: null, href: null, linkClass: 'navbar-separator' },
-            { name: 'Sign In', href: '#' },
-            { name: 'Sign Up', href: '#' },
-
-        ]
-};
+import { sudokuNavBarLinks } from "../commons/navbarLinks";
 
 class SudokuHome extends React.Component {
     render(): Array<JSX.Element> {
